cleaning up loading of initial image
